test:add ULogoutButton golden test and refactor the tests for buttons
