+++ packages/ui_library/test/widgets/buttons/u_button/goldens/u_button_secondary.png
