+++ packages/ui_library/test/widgets/buttons/u_fab/goldens/u_fab_big.png
