+++ packages/ui_library/test/widgets/buttons/u_icon_button/goldens/u_icon_button_search.png
